types

+++ beacon/builder/service.go
@@ -93,7 +93,8 @@ func (s *Service) RequestBestBlock(
 	}
 
 	parentBlockRoot, err := st.GetBlockRootAtIndex(
-		(st.GetSlot() - 1) % s.BeaconCfg().SlotsPerHistoricalRoot)
+		uint64(st.GetSlot()-1) % s.BeaconCfg().SlotsPerHistoricalRoot,
+	)
 	if err != nil {
 		return nil, nil, err
 	}
